Modify operation so that forums with many topics are split into 2 or more map files.  This avoids the external limit of 50000 links per xml file, and significantly reduces the sql query time when the sitemap for such a forum is generated.
Also made the generation of sitemaps for viewforum pages optional
(default false) as these do not usually contain content that is
beneficial to the purpose of a sitemap.

Also removed the caching of sql queries, as on reflection, the
benfits are not worthwhile, especially after breaking up the
queries for large forums.